Refactored Providers to use DI.
Added domain specific extension
interfaces for Scanners : MeetupScanner , AlleventsScanner ,
EventbriteScanner to be able to slot implementations per domain into the DI container.

Created WebProxyDriver factory to get instances for scanners. Scanners
dispose of there WebDriver after use so the Dependency should be on a
factory not an instance of WebProxyDriver.